fixed refactored code, added latset changes, tested
