+++ server/routes/auth/index.js
@@ -1,7 +1,7 @@
 'use strict';
 
 var express = require('express');
-var controller = require('./form.controller');
+var controller = require('./auth.controller');
 
 var router = express.Router();
